Fix "YOU DIED" instead of "THE END"
If you get an ending, then get an ending again without ever pausing in between, the "THE END" screen will hilariously be a "YOU DIED" screen instead. The issue is that the function NullGame sets AchievementsMenu (the value that represents whether you're in the achievements menu) to -1, but DrawEnding checks if AchievementsMenu=0 to draw the "THE END" screen. So if you never paused to reset your AchievementsMenu value to 0, it will remain at -1 and the "THE END" screen won't be drawn. Fixed by replacing = with <=.